+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Change testing framework for modules and subworkflows from pytest to nf-test ([#2490](https://github.com/nf-core/tools/pull/2490))
 - `bump_version` keeps now the indentation level of the updated version entries ([#2514](https://github.com/nf-core/tools/pull/2514))
 - Run tests with Python 3.12 ([#2522](https://github.com/nf-core/tools/pull/2522)).
+- Use Path objects for ComponentCreate and update the structure of components templates.
 
 # [v2.10 - Nickel Ostrich](https://github.com/nf-core/tools/releases/tag/2.10) + [2023-09-25]